requirements.txt
# FastAPI and web framework
fastapi==0.120.4
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.1

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-dotenv==1.2.1
argon2_cffi

# Pydantic
pydantic==2.12.3
pydantic-settings==2.11.0
email-validator==2.3.0

# AWS S3
boto3==1.40.66

# HTTP requests
requests==2.32.5
apscheduler==3.10.4
